+++ web/src/app/schedules/on-call-notifications/ScheduleOnCallNotificationsList.tsx
@@ -5,7 +5,8 @@ import Avatar from '@material-ui/core/Avatar'
 import FlatList from '../../lists/FlatList'
 import OtherActions from '../../util/OtherActions'
 import { SlackBW } from '../../icons/components/Icons'
-import { useRulesData, ruleSummary } from './util'
+import { useRulesData } from './hooks'
+import { ruleSummary } from './util'
 import ScheduleOnCallNotificationsCreateDialog from './ScheduleOnCallNotificationsCreateDialog'
 import ScheduleOnCallNotificationsDeleteDialog from './ScheduleOnCallNotificationsDeleteDialog'
 import CreateFAB from '../../lists/CreateFAB'
